Finished Turret, Began Ekkos Watch
Completed the turret, turned it into an enemy now spawnable by the T-800 boss, still need to test it. Began Ekkos time watch ability. It will spawn an object that after 4 seconds, teleports the player back to it and sets their health back to whatever it was when they used the ability first