The Grand Finale
Well, its been a journey. Through tears of pride, joy and pain I have created a master piece of video game entertainment. Im going to stop adding more stuff until I hand it in, I might just keep fucking with it and submit it in 2025 for the madhouse as a surprise release with a story and more levels etc, actually does seem quite fun. Anyway id like to thank the academy, my family and god... Now ive just got to trawl through the git commits and retype half of the NEA in a week, should be easy...

Sam out